+++ src/palettepage.cpp
@@ -15,17 +15,19 @@
 PalettePage::PalettePage(QWidget* parent)
     : QWidget(parent),
       mColorScheme(6, QColor(0, 255, 0)),
-      mArduinoPaletteScrollArea{new PaletteScrollArea(this)},
-      mHuePaletteScrollArea{new PaletteScrollArea(this)},
+      mPaletteScrollArea{new PaletteScrollArea(this)},
       mColorPicker{new MultiColorPicker(this)} {
     grabGesture(Qt::SwipeGesture);
 
-    mArduinoPaletteScrollArea->setupButtons(true);
-    mHuePaletteScrollArea->setupButtons(false);
+    connect(mPaletteScrollArea,
+            SIGNAL(paletteClicked(EPalette)),
+            this,
+            SLOT(paletteButtonClicked(EPalette)));
+
     mColorPicker->setVisible(false);
 
-    mMode = EGroupMode::arduinoPresets;
-    setMode(EGroupMode::HSV);
+    mMode = EGroupMode::presets;
+    setMode(EGroupMode::wheel);
 }
 
 
@@ -33,20 +35,16 @@ PalettePage::PalettePage(QWidget* parent)
 // Slots
 // ----------------------------
 
-void PalettePage::paletteButtonClicked(ERoutine routine, EPalette palette) {
+void PalettePage::paletteButtonClicked(EPalette palette) {
     mPaletteEnum = palette;
     mColorScheme = mPresetPalettes.palette(mPaletteEnum).colors();
-    if (mMode == EGroupMode::arduinoPresets) {
-        mArduinoPaletteScrollArea->highlightRoutineButton(routine, mPaletteEnum);
-    } else if (mMode == EGroupMode::huePresets) {
-        mHuePaletteScrollArea->highlightRoutineButton(routine, mPaletteEnum);
-    }
+    mPaletteScrollArea->highlightButton(mPaletteEnum);
 
-    emit routineUpdate(routine, mPaletteEnum);
+    emit paletteUpdate(palette);
 }
 
 void PalettePage::updateBrightness(std::uint32_t brightness) {
-    if (mMode == EGroupMode::HSV) {
+    if (mMode == EGroupMode::wheel) {
         auto color = mColorScheme[mColorPicker->selectedLight()];
         color.setHsvF(color.hueF(), color.saturationF(), color.valueF() / 100.0);
         mColorScheme[mColorPicker->selectedLight()] = color;
@@ -59,10 +57,7 @@ void PalettePage::updateBrightness(std::uint32_t brightness) {
 // ----------------------------
 
 
-void PalettePage::update(std::size_t count,
-                         const std::vector<QColor>& colorScheme,
-                         bool hasArduinoDevices,
-                         bool hasNanoleafDevices) {
+void PalettePage::update(std::size_t count, const std::vector<QColor>& colorScheme) {
     if (mColorScheme.empty()) {
         if (count > 0) {
             mColorPicker->updateBrightness(colorScheme[0].valueF() * 100.0);
@@ -73,30 +68,17 @@ void PalettePage::update(std::size_t count,
     mColorScheme = colorScheme;
     lightCountChanged(count);
     mColorPicker->updateColorStates(mColorScheme);
-    if (mMode == EGroupMode::huePresets && (hasArduinoDevices || hasNanoleafDevices)) {
-        setMode(EGroupMode::arduinoPresets);
-    } else if (mMode == EGroupMode::arduinoPresets && !(hasArduinoDevices || hasNanoleafDevices)) {
-        setMode(EGroupMode::huePresets);
-    }
-}
-
-void PalettePage::newRoutineSelected(ERoutine routine) {
-    emit routineUpdate(routine, mPaletteEnum);
 }
 
 void PalettePage::setMode(EGroupMode mode) {
     if (mMode != mode) {
-        mArduinoPaletteScrollArea->setVisible(false);
-        mHuePaletteScrollArea->setVisible(false);
+        mPaletteScrollArea->setVisible(false);
         mColorPicker->setVisible(false);
         switch (mode) {
-            case EGroupMode::arduinoPresets:
-                mArduinoPaletteScrollArea->setVisible(true);
+            case EGroupMode::presets:
+                mPaletteScrollArea->setVisible(true);
                 break;
-            case EGroupMode::huePresets:
-                mHuePaletteScrollArea->setVisible(true);
-                break;
-            case EGroupMode::HSV:
+            case EGroupMode::wheel:
                 mPaletteEnum = EPalette::custom;
                 mColorPicker->setVisible(true);
                 break;
@@ -108,32 +90,20 @@ void PalettePage::setMode(EGroupMode mode) {
 void PalettePage::resize() {
     auto yPos = int(height() * 0.05);
     QSize scrollAreaSize(int(width()), int(height() * 0.94f));
-    mArduinoPaletteScrollArea->setGeometry(0,
-                                           yPos,
-                                           scrollAreaSize.width(),
-                                           scrollAreaSize.height());
-    mArduinoPaletteScrollArea->resize();
-
-    mHuePaletteScrollArea->setGeometry(0, yPos, scrollAreaSize.width(), scrollAreaSize.height());
-    mHuePaletteScrollArea->resize();
+    mPaletteScrollArea->setGeometry(0, yPos, scrollAreaSize.width(), scrollAreaSize.height());
+    mPaletteScrollArea->resize();
 
     mColorPicker->setGeometry(0, yPos, width(), int(height() * 0.94));
     mColorPicker->resize();
 }
 
-void PalettePage::showEvent(QShowEvent*) {
-    resize();
-}
-
 void PalettePage::lightCountChanged(std::size_t count) {
     mColorPicker->updateColorCount(count);
     if (count == 0) {
-        mArduinoPaletteScrollArea->setEnabled(false);
-        mHuePaletteScrollArea->setEnabled(false);
+        mPaletteScrollArea->setEnabled(false);
         mColorPicker->enable(false, EColorPickerType::color);
     } else {
-        mArduinoPaletteScrollArea->setEnabled(true);
-        mHuePaletteScrollArea->setEnabled(true);
+        mPaletteScrollArea->setEnabled(true);
         mColorPicker->enable(true, EColorPickerType::color);
     }
 }